arch2/metadata/requirements.txt
flask
grpcio==1.75.1
grpcio-tools==1.75.1
protobuf==6.32.1